Move rescue board components to their own subdir.
