+++ src/js/components/RescueBoard/RatList.jsx
@@ -1,7 +1,7 @@
 // App imports
-import Component from './Component.jsx';
+import Component from '../Component.jsx';
 import Rat from './Rat.jsx';
-import AssignedRatPropType from '../types/AssignedRat.js';
+import AssignedRatPropType from '../../types/AssignedRat.js';
 
 // Module imports
 import React from 'react';
